refactor: extract GlobalFilter class to a separate file
