+++ tests/Query.test.ts
@@ -5,7 +5,8 @@ import moment from 'moment';
 import { Query } from '../src/Query/Query';
 import { Status } from '../src/Status';
 import { Priority, Task } from '../src/Task';
-import { GlobalFilter, getGlobalFilter, resetSettings, updateSettings } from '../src/Config/Settings';
+import { resetSettings, updateSettings } from '../src/Config/Settings';
+import { GlobalFilter, getGlobalFilter } from '../src/Config/GlobalFilter';
 import { TaskLocation } from '../src/TaskLocation';
 import { createTasksFromMarkdown, fromLine } from './TestHelpers';
 import { shouldSupportFiltering } from './TestingTools/FilterTestHelpers';
