+++ tests/Query/Filter/DescriptionField.test.ts
@@ -3,7 +3,8 @@
  */
 import moment from 'moment';
 import { DescriptionField } from '../../../src/Query/Filter/DescriptionField';
-import { GlobalFilter, resetSettings, updateSettings } from '../../../src/Config/Settings';
+import { resetSettings, updateSettings } from '../../../src/Config/Settings';
+import { GlobalFilter } from '../../../src/Config/GlobalFilter';
 import { testTaskFilter } from '../../TestingTools/FilterTestHelpers';
 import { fromLine } from '../../TestHelpers';
 import type { FilterOrErrorMessage } from '../../../src/Query/Filter/Filter';
