+++ tests/Task.test.ts
@@ -5,7 +5,8 @@ import moment from 'moment';
 import type { Moment } from 'moment';
 import { Status } from '../src/Status';
 import { Priority, Task } from '../src/Task';
-import { GlobalFilter, resetSettings, updateSettings } from '../src/Config/Settings';
+import { resetSettings, updateSettings } from '../src/Config/Settings';
+import { GlobalFilter } from '../src/Config/GlobalFilter';
 import type { StatusCollection } from '../src/StatusCollection';
 import { StatusRegistry } from '../src/StatusRegistry';
 import { TaskLocation } from '../src/TaskLocation';
